CLN: Remove unused funcs
